cpufreq: Add more governors
* SmartassН3 
(Based on SmartassV2 with parameters tuned by H3ROS. It has been tuned
to have more efficient frequency scaling (not ramping too high when not
needed) which this could potentially result in better battery life.)
* Badass
(Badass removes all of this "fast peaking" to the max frequency. To
trigger a frequency increase, the system must run a bit with high load,
then the frequency is bumped. If that is still not enough the governor
gives you full throttle. (this transition should not take longer than
1-2 seconds, depending on the load your system is experiencing)
Badass will also take the gpu load into consideration. If the gpu is
moderately busy it will bypass the above check and clock the cpu to max
frequency. If the gpu is crushed under load, badass will lift the
restrictions to the cpu.)
* Hyper
(The Hyper (formerly known as kenobi) is an aggressive smart and smooth
governor based on the Ondemand and is equipped with several features of
Ondemandx suspend profiles. It also has the fast_start deep_sleep
variable and detection features. In addition, the maximum frequency is
in suspend mode 500Mhz or whatever the kernel devel. This is a more
smoothness oriented governor which means that it is good for
performance, without sacrificing much battery life. )
* Wheatley
(This governor is build on “ondemand” but increases the C4 (the sleep
state) state time of the CPU and doing so trying to save juice. So the
results show that Wheatley works as intended and ensures that the C4
state is used whenever the task allows a proper efficient usage of the
C4 state. For more demanding tasks which cause a large number of wakeups
and prevent the efficient usage of the C4 state, the governor resorts to
the next best power saving mechanism and scales down the frequency. So
with the new highly-flexible Wheatley governor one can have the best of
both worlds. Obviously, this governor is only available on multi-core
devices.
Wheatley is a more performance orientated governor as it scales more
aggressively than ondemand and sticks with higher frequencies.)